backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
